+++ src/pro/main/ipc/handlers/local_agent/local_agent_handler.ts
@@ -1118,6 +1118,7 @@ export async function handleLocalAgentStream(
             const closingThinkBlock = "</think>\n";
             fullResponse += closingThinkBlock;
             await updateResponseInDb(placeholderMessageId, fullResponse);
+            sendChunk(fullResponse);
           }
           activeRetryReplayEvents = null;
